move AR length validation tests into separate test-case.
